+++ custats/backend/functions.h
@@ -6,9 +6,9 @@
 #include "common.h"
 #include "kernels.h"
 
-void poisson_logpmf_hh2h(const int *k, const double *r, double *out, const int size);
-void poisson_logpmf_dh2d(const int *k, const double *r, double *out, const int size);
-void poisson_logpmf_hd2d(const int *k, const double *r, double *out, const int size);
-void poisson_logpmf_dd2d(const int *k, const double *r, double *out, const int size);
+void poisson_logpmf_hh2h(const int *k, const float *r, float *out, const int size);
+void poisson_logpmf_dh2d(const int *k, const float *r, float *out, const int size);
+void poisson_logpmf_hd2d(const int *k, const float *r, float *out, const int size);
+void poisson_logpmf_dd2d(const int *k, const float *r, float *out, const int size);
 
 #endif // FUNCTIONS_H_
